# rewrite eventsByPersistenceId as a stage