--- a/build.sbt
+++ b/build.sbt
@@ -52,8 +52,13 @@ def common: Seq[Setting[_]] = SbtScalariform.scalariformSettings ++ Seq(
 
   releaseCrossBuild := true,
 
+  logBuffered in Test := System.getProperty("akka.logBufferedTests", "false").toBoolean,
+
   // show full stack traces and test case durations
   testOptions in Test += Tests.Argument("-oDF"),
+
+  // don't save test output to a file
+  testListeners in (Test, test) := Seq(TestLogger(streams.value.log, {_ => streams.value.log }, logBuffered.value)),
 
   // -v Log "test run started" / "test started" / "test run finished" events on log level "info" instead of "debug".
   // -a Show stack traces and exception class name for AssertionErrors.