--- a/core/src/main/resources/reference.conf
+++ b/core/src/main/resources/reference.conf
@@ -546,6 +546,13 @@ cassandra-query-journal {
   # Value less or equal to 0 means max-result-size will be used
   # http://docs.datastax.com/en/drivers/java/3.0/com/datastax/driver/core/Statement.html
   max-result-size-query = 250
+  
+  # When the available rows without fetching in the ResultSet reach this threshold
+  # more will be fetched asynchronously. 
+  # Value should be between 0.0 and 1.0.
+  # E.g. threshold of 0.1 and fetch size of 250 will trigger new fetch when 
+  # 25 rows are available. 
+  fetch-more-threshold = 0.1
 
   # Read consistency level
   read-consistency = "QUORUM"