--- a/core/src/main/resources/reference.conf
+++ b/core/src/main/resources/reference.conf
@@ -217,10 +217,10 @@ cassandra-journal {
   target-partition-size = 500000
 
   # Maximum size of result set
-  max-result-size = 50001
+  max-result-size = 250
 
   # Maximum size of result set during replay
-  max-result-size-replay = 50001
+  max-result-size-replay = 250
 
   # The query journal to use when recovering
   query-plugin = "cassandra-query-journal"