--- a/core/src/main/resources/reference.conf
+++ b/core/src/main/resources/reference.conf
@@ -537,6 +537,12 @@ cassandra-query-journal {
 
   # New events are retrieved (polled) with this interval.
   refresh-interval = 3s
+  
+  # Sequence numbers for a persistenceId is assumed to be monotonically increasing
+  # without gaps. That is used for detecting missing events.
+  # In early versions of the journal that might not be true and therefore
+  # this can be relaxed by setting this property to off.
+  gap-free-sequence-numbers = on
 
   # How many events to fetch in one query (replay) and keep buffered until they
   # are delivered downstreams.
@@ -552,7 +558,7 @@ cassandra-query-journal {
   # Value should be between 0.0 and 1.0.
   # E.g. threshold of 0.1 and fetch size of 250 will trigger new fetch when 
   # 25 rows are available. 
-  fetch-more-threshold = 0.1
+  fetch-more-threshold = 0.2
 
   # Read consistency level
   read-consistency = "QUORUM"