--- a/core/src/main/scala/akka/persistence/cassandra/query/CassandraReadJournalConfig.scala
+++ b/core/src/main/scala/akka/persistence/cassandra/query/CassandraReadJournalConfig.scala
@@ -19,6 +19,11 @@ import akka.annotation.InternalApi
   val refreshInterval: FiniteDuration = config.getDuration("refresh-interval", MILLISECONDS).millis
   val maxBufferSize: Int = config.getInt("max-buffer-size")
   val fetchSize: Int = config.getInt("max-result-size-query")
+  val fetchMoreThreshold: Double = config.getDouble("fetch-more-threshold")
+  require(
+    0.0 <= fetchMoreThreshold && fetchMoreThreshold <= 1.0,
+    s"fetch-more-threshold must be between 0.0 and 1.0, was $fetchMoreThreshold"
+  )
   val readConsistency: ConsistencyLevel = ConsistencyLevel.valueOf(config.getString("read-consistency"))
   val readRetries: Int = config.getInt("read-retries")
   val firstTimeBucket: TimeBucket = TimeBucket(config.getString("first-time-bucket"))