… core/src/main/scala/akka/persistence/cassandra/query/EventsByPersistenceIdPublisher.scala
… core/src/test/scala/akka/persistence/cassandra/query/EventsByPersistenceIdSpec.scala
… core/src/test/scala/akka/persistence/cassandra/query/EventsByPersistenceIdWithControlSpec.scala
… core/src/test/scala/akka/persistence/cassandra/query/TestActor.scala