Here is a rewrite of the EventsByPersistenceIdPublisher as a GraphStage. There are a few new things I want to add, but this is good for first review.