@patriknw 2017-10-04
@chbatey I have update this with your feedback, and the additional features for poll and fast-forward. I want to run some perf tests also, but should be ready for another round of review.