@patriknw 2017-10-05
Did a sanity check of the performance, comparing with master. I don't trust the exact results of this test but I think it's good enough to ensure that we are not introducing a regression.

Tried with a few eventCount and fetch sizes:
```
testOnly akka.persistence.cassandra.journal.CassandraJournalPerfSpec -- -z recovering

eventCount: 100,  iterations: 100, fetch-size: 250
  new avg: 10 ms
  old avg: 11 ms

eventCount: 1000, iterations: 100, fetch-size: 100
  new avg: 37 ms
  old avg: 31 ms
fetch-size: 250
  new avg: 30 ms
  old avg: 25 ms
fetch-size: 1000
  new avg: 28 ms
  old avg: 24 ms

eventCount: 10000, iterations: 100, fetch-size: 100
  new avg: 212 ms
  old avg: 185 ms
fetch-size: 250
  new avg: 194 ms
  old avg: 184 ms
fetch-size: 1000
  new avg: 164 ms
  old avg: 154 ms
```