@patriknw 2017-10-09
@chbatey I also added sequence number gap detection. Take a look at a367a68d37ca4a2439673d745f82d9c2a51b82fa, then we can release new version.